requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
flask-mail==0.9.1
python-dotenv==1.0.1
flask-cors==4.0.0
requests==2.31.0
gunicorn==21.2.0 